# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DentalAI SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dental/appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: date
          in: query
          schema:
            type: string
            description: YYYY-MM-DD filter date
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentWithPatient"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "409":
          description: Conflict - time slot unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
  /dental/appointments/conflicts:
    get:
      operationId: checkAppointmentConflicts
      tags: [appointments]
      summary: Check for conflicts in a proposed time window
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: startsAt
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endsAt
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: excludeAppointmentId
          in: query
          schema:
            type: integer
            description: Exclude this appointment ID when checking (for reschedules)
      responses:
        "200":
          description: Conflict check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictCheckResponse"
  /dental/appointments/availability:
    get:
      operationId: getAvailability
      tags: [appointments]
      summary: Get available time slots for a date
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: date
          in: query
          required: true
          schema:
            type: string
            description: YYYY-MM-DD
        - name: durationMinutes
          in: query
          schema:
            type: integer
            default: 60
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
  /dental/appointments/{appointmentId}:
    parameters:
      - $ref: "#/components/parameters/TenantHeader"
      - name: appointmentId
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithPatient"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel/delete an appointment
      responses:
        "204":
          description: Deleted

  # ──────────────────────────── LEADS ────────────────────────────
...
  /dental/reports/appointments:
    get:
      operationId: getAppointmentsReport
      tags: [reports]
      summary: Appointments by period report
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: day
      responses:
        "200":
          description: Appointments report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentsReport"
...
components:
  parameters:
...
    TenantHeader:
      name: x-tenant-id
      in: header
      required: true
      schema:
        type: integer
      description: Tenant (clinic) ID
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        birthDate:
          type: string
          nullable: true
        cpf:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        profilePicUrl:
          type: string
          nullable: true
        firstVisit:
          type: string
          format: date-time
          nullable: true
        lastVisit:
          type: string
          format: date-time
          nullable: true
        totalSpent:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - phone
        - totalSpent
        - createdAt
        - updatedAt
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        patientId:
          type: integer
          nullable: true
        leadId:
          type: integer
          nullable: true
        procedureId:
          type: integer
          nullable: true
        procedureName:
          type: string
          nullable: true
        status:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        price:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        confirmed:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - status
        - startsAt
        - endsAt
        - confirmed
        - createdAt
        - updatedAt
    AppointmentWithPatient:
      allOf:
        - $ref: "#/components/schemas/Appointment"
        - type: object
          properties:
            patient:
              $ref: "#/components/schemas/Patient"
    CreateAppointmentBody:
      type: object
      properties:
        patientId:
          type: integer
          nullable: true
        leadId:
          type: integer
          nullable: true
        procedureId:
          type: integer
          nullable: true
        procedureName:
          type: string
          nullable: true
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        price:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - startsAt
        - endsAt
    UpdateAppointmentBody:
      type: object
      properties:
        patientId:
          type: integer
        status:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        price:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        confirmed:
          type: string
        procedureId:
          type: integer
          nullable: true
        procedureName:
          type: string
          nullable: true
    AvailabilityResponse:
      type: object
      properties:
        date:
          type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
      required:
        - date
        - slots
    TimeSlot:
      type: object
      properties:
        start:
          type: string
        end:
          type: string
        available:
          type: boolean
      required:
        - start
        - end
        - available
    ConflictCheckResponse:
      type: object
      properties:
        hasConflict:
          type: boolean
        conflicts:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentWithPatient"
      required:
        - hasConflict
        - conflicts
    ConflictError:
      type: object
      properties:
        error:
          type: string
        conflictingAppointment:
          $ref: "#/components/schemas/Appointment"
      required:
        - error

    # ─── LEAD ───
...
    AppointmentsReport:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              period:
                type: string
              count:
                type: integer
              completed:
                type: integer
              cancelled:
                type: integer
            required:
              - period
              - count
              - completed
              - cancelled
        total:
          type: integer
        completed:
          type: integer
        cancelled:
          type: integer
      required:
        - data
        - total
        - completed
        - cancelled
...